added vira9410

+++ run_viraocr.sh
@@ -15,11 +15,13 @@ mkdir -p $basepath/runs/$runname
 for dataset in simple medium difficult ; do
 	for absfilepath in $basepath/benchmark_data/$dataset/*.{jpg,png,bmp,tiff} ; do
 		if [ -e $absfilepath ]; then
-			echo Running on $absfilepath
 			filename=$(basename "$absfilepath")
 			basefilename="${filename%.*}"
-			matlab -nojvm -nodesktop -nosplash -r "cd /opt/farsi-ocr-engine-v3; setup; init; ocr_main_v3('$absfilepath', true, true, true); exit"
-			cp /opt/farsi-ocr-engine-v3/temp/$basefilename/output.txt $basepath/runs/$runname/${basefilename}.txt
+			if [ -e $basepath/benchmark_data/$dataset/${basefilename}.txt ]; then
+				echo Running on $absfilepath
+				matlab -nojvm -nodesktop -nosplash -r "cd /opt/farsi-ocr-engine-v3; setup; init; ocr_main_v3('$absfilepath', false, true, true); exit"
+				cp /opt/farsi-ocr-engine-v3/temp/$basefilename/output.txt $basepath/runs/$runname/${basefilename}.txt
+			fi
 		fi
 	done
 done
